Don't error out if the input isn't among the filters
If the input URL wasn't twitter.com or tiktok.com the script ended up
erroring out; and for a good reason:
it was feeding the reply function a non existent variable because the
code straight up ignored anything that wasn't among the matched URLs.

And use `#` instead of quotes for the block comment.